requirements.txt
fastapi
mongoengine
uvicorn
pyyaml
